requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy
opencv-python
picamera2
libcamera
pyserial
pygame
